requirements.txt
openai>=1.12.0
python-dotenv>=1.0.0
numpy>=1.26.0
scipy>=1.11.0
tqdm>=4.66.0
ipdb>=0.13.13
loguru>=0.7.0
pillow>=9.5.0
